+++ include/llst.h
@@ -6,7 +6,7 @@
 /*   By: ll-hotel <ll-hotel@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/17 19:35:45 by ll-hotel          #+#    #+#             */
-/*   Updated: 2024/04/17 22:54:23 by ll-hotel         ###   ########.fr       */
+/*   Updated: 2024/04/20 17:15:36 by ll-hotel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,6 +28,7 @@ struct	s_llst
 
 void	llst_addfront(t_llst_head *lst, t_llst *element);
 void	llst_addback(t_llst_head *lst, t_llst *element);
+void	*llst_get_last(t_llst_head *lst);
 void	llst_clear(t_llst_head *lst, void (*f)(void *));
 long	llst_len(t_llst_head *lst);
