+++ src/parser.c
@@ -6,34 +6,46 @@
 /*   By: ll-hotel <ll-hotel@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/19 15:59:10 by ll-hotel          #+#    #+#             */
-/*   Updated: 2024/04/19 20:41:27 by ll-hotel         ###   ########.fr       */
+/*   Updated: 2024/04/23 23:10:38 by ll-hotel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-static void	*parser_insert_env_var(t_token *token_dollar, t_env *env);
+static void	*parser_insert_env_var(t_token *name, t_env *env);
 
-void	*parser(t_llst *token_lst, t_env *env)
+static void	move_head_forward(t_llst_head *head, t_token **token)
 {
-	t_token	*token;
+	void	*tmp;
 
-	token = (t_token *)token_lst;
+	tmp = token[0]->next;
+	free(*token);
+	head->first = tmp;
+	*token = tmp;
+}
+
+void	*parser(t_llst_head *token_lst, t_env *env)
+{
+	t_llst_head	head;
+	t_token		*token;
+
+	head.first = token_lst->first;
+	token = (t_token *)head.first;
 	while (token)
 	{
 		if (token->type == TOKEN_SIMPLE_QUOTE)
 		{
-			token = token->next;
+			move_head_forward(&head, &token);
 			while (token && token->type != TOKEN_SIMPLE_QUOTE)
 				token = token->next;
 		}
 		else if (token->type == TOKEN_DOUBLE_QUOTE)
 		{
 			token = token->next;
-			while (token && token->type != TOKEN_SIMPLE_QUOTE)
+			while (token && token->type != TOKEN_DOUBLE_QUOTE)
 			{
 				if (token->type == TOKEN_DOLLAR)
-					token = parser_insert_env_var(token->next, env);
+					token = parser_insert_env_var(token, env);
 				if (token)
 					token = token->next;
 			}
@@ -44,50 +56,15 @@ void	*parser(t_llst *token_lst, t_env *env)
 	return (token_lst);
 }
 
-char	*parser_assemble(t_token *token)
+static void	*parser_insert_env_var(t_token *dollar, t_env *env)
 {
-	char	*str;
-	void	*joined;
+	t_token		*token_var;
 
-	str = ft_calloc(1, sizeof(*str));
-	if (!str)
-		return (NULL);
-	while (token)
+	(void)env;
+	token_var = dollar->next;
+	if (token_var->type != TOKEN_WORD || token_var->str == NULL)
 	{
-		joined = NULL;
-		if (token->type == TOKEN_WORD)
-			joined = ft_strjoin(str, token->str);
-		else if (token->type == TOKEN_REDIRECT)
-			joined = ft_strjoin(str, token->str);
-		else if (token->type == TOKEN_SPACE)
-			joined = ft_strjoin(str, " ");
-		else if (token->type == TOKEN_DOLLAR)
-			joined = ft_strjoin(str, "$");
-		else if (token->type == TOKEN_SIMPLE_QUOTE)
-			joined = ft_strjoin(str, "\'");
-		else if (token->type == TOKEN_DOUBLE_QUOTE)
-			joined = ft_strjoin(str, "\"");
-		else if (token->type == TOKEN_PIPE)
-			joined = ft_strjoin(str, "|");
-		free(str);
-		if (!joined)
-			return (NULL);
-		str = joined;
-		token = token->next;
+		;
 	}
-	return (str);
-}
-
-static void	*parser_insert_env_var(t_token *token_dollar, t_env *env)
-{
-	t_token		*token;
-	t_env_var	*var;
-
-	token = token_dollar->next;
-	if (token->type != TOKEN_WORD || token->str == NULL)
-		return (token);
-	var = env_var_get(env, token->str);
-	if (!var)
-		return (token);
-	return (token);
+	return (NULL);
 }
